Add support for loadbalancer interface
This adds support for the `loadbalancer` interface so that cloud-native
LBs can be provided by the integrator charms. Additionally, it
simplifies the confusing way the relations between the masters and
workers change depending on whether kubeapi-load-balancer is being used
or not by making that use the same `lb-provider` endpoint and always
forwarding the API endpoint URLs via the `kube-control` relation.

Part of [lp:1897818][]
Depends on:
  * https://github.com/juju-solutions/loadbalancer-interface/pull/13
  * https://github.com/juju-solutions/interface-kube-control/pull/33
  * https://github.com/charmed-kubernetes/charm-kubernetes-worker/pull/84
  * https://github.com/charmed-kubernetes/charm-kubeapi-load-balancer/pull/11

[lp:1897818]: https://bugs.launchpad.net/charmed-kubernetes-testing/+bug/1897818